Allow graph node scenes to be selected and added to graph edit
The selection pane now has an Add Scene button that allows the user to select a graph edit scene file and create a button that will add it to the graph edit. Once the new button is pressed, the graph node is added to the graph edit and is free to be manipulated.